split element.rs
