+++ crates/ui/src/input/input.rs
@@ -4,28 +4,28 @@
 //! https://github.com/zed-industries/zed/blob/main/crates/gpui/examples/input.rs
 
 use std::ops::Range;
+use unicode_segmentation::*;
+
+use gpui::prelude::FluentBuilder as _;
+use gpui::{
+    actions, div, point, px, AnyElement, AppContext, Bounds, ClickEvent, ClipboardItem,
+    Context as _, EventEmitter, FocusHandle, FocusableView, InteractiveElement as _, IntoElement,
+    KeyBinding, KeyDownEvent, Model, MouseButton, MouseDownEvent, MouseMoveEvent, MouseUpEvent,
+    ParentElement as _, Pixels, Point, Rems, Render, ShapedLine, SharedString, Styled as _,
+    UTF16Selection, ViewContext, ViewInputHandler, WindowContext,
+};
 
 use super::blink_cursor::BlinkCursor;
 use super::change::Change;
+use super::element::TextElement;
 use super::ClearButton;
+
 use crate::history::History;
 use crate::indicator::Indicator;
 use crate::theme::ActiveTheme;
 use crate::StyledExt as _;
 use crate::{event::InteractiveElementExt as _, Size};
 use crate::{Sizable, StyleSized};
-use gpui::prelude::FluentBuilder as _;
-use gpui::{
-    actions, div, fill, point, px, relative, rems, size, AnyElement, AppContext, Bounds,
-    ClickEvent, ClipboardItem, Context as _, Element, ElementId, ElementInputHandler, EventEmitter,
-    FocusHandle, FocusableView, GlobalElementId, InteractiveElement as _, IntoElement, KeyBinding,
-    KeyDownEvent, LayoutId, Model, MouseButton, MouseDownEvent, MouseMoveEvent, MouseUpEvent,
-    PaintQuad, ParentElement as _, Pixels, Point, Rems, Render, ShapedLine, SharedString, Style,
-    Styled as _, TextRun, UTF16Selection, UnderlineStyle, View, ViewContext, ViewInputHandler,
-    WindowContext, WrappedLine,
-};
-use smallvec::SmallVec;
-use unicode_segmentation::*;
 
 actions!(
     input,
@@ -119,29 +119,29 @@ pub fn init(cx: &mut AppContext) {
 }
 
 pub struct TextInput {
-    focus_handle: FocusHandle,
-    text: SharedString,
-    multi_line: bool,
-    history: History<Change>,
-    blink_cursor: Model<BlinkCursor>,
-    prefix: Option<Box<dyn Fn(&mut ViewContext<Self>) -> AnyElement + 'static>>,
-    suffix: Option<Box<dyn Fn(&mut ViewContext<Self>) -> AnyElement + 'static>>,
-    loading: bool,
-    placeholder: SharedString,
-    selected_range: Range<usize>,
-    selection_reversed: bool,
-    marked_range: Option<Range<usize>>,
-    last_layout: Option<ShapedLine>,
-    last_bounds: Option<Bounds<Pixels>>,
-    last_cursor_offset: Option<usize>,
-    last_selected_range: Option<Range<usize>>,
-    scroll_offset: Point<Pixels>,
-    is_selecting: bool,
-    disabled: bool,
-    masked: bool,
-    appearance: bool,
-    cleanable: bool,
-    size: Size,
+    pub(super) focus_handle: FocusHandle,
+    pub(super) text: SharedString,
+    pub(super) multi_line: bool,
+    pub(super) history: History<Change>,
+    pub(super) blink_cursor: Model<BlinkCursor>,
+    pub(super) prefix: Option<Box<dyn Fn(&mut ViewContext<Self>) -> AnyElement + 'static>>,
+    pub(super) suffix: Option<Box<dyn Fn(&mut ViewContext<Self>) -> AnyElement + 'static>>,
+    pub(super) loading: bool,
+    pub(super) placeholder: SharedString,
+    pub(super) selected_range: Range<usize>,
+    pub(super) selection_reversed: bool,
+    pub(super) marked_range: Option<Range<usize>>,
+    pub(super) last_layout: Option<ShapedLine>,
+    pub(super) last_bounds: Option<Bounds<Pixels>>,
+    pub(super) last_cursor_offset: Option<usize>,
+    pub(super) last_selected_range: Option<Range<usize>>,
+    pub(super) scroll_offset: Point<Pixels>,
+    pub(super) is_selecting: bool,
+    pub(super) disabled: bool,
+    pub(super) masked: bool,
+    pub(super) appearance: bool,
+    pub(super) cleanable: bool,
+    pub(super) size: Size,
     pattern: Option<regex::Regex>,
     validate: Option<Box<dyn Fn(&str) -> bool + 'static>>,
 }
@@ -513,7 +513,7 @@ impl TextInput {
         cx.notify()
     }
 
-    fn cursor_offset(&self) -> usize {
+    pub(super) fn cursor_offset(&self) -> usize {
         if self.selection_reversed {
             self.selected_range.start
         } else {
@@ -679,7 +679,7 @@ impl TextInput {
         self.pause_blink_cursor(cx)
     }
 
-    fn on_drag_move(&mut self, event: &MouseMoveEvent, cx: &mut ViewContext<Self>) {
+    pub(super) fn on_drag_move(&mut self, event: &MouseMoveEvent, cx: &mut ViewContext<Self>) {
         if self.text.is_empty() {
             return;
         }
@@ -870,280 +870,6 @@ impl FocusableView for TextInput {
     }
 }
 
-struct TextElement {
-    input: View<TextInput>,
-}
-impl TextElement {
-    fn paint_mouse_listeners(&mut self, cx: &mut WindowContext) {
-        cx.on_mouse_event({
-            let input = self.input.clone();
-
-            move |event: &MouseMoveEvent, _, cx| {
-                if event.pressed_button == Some(MouseButton::Left) {
-                    input.update(cx, |input, cx| {
-                        input.on_drag_move(event, cx);
-                    });
-                }
-            }
-        });
-    }
-}
-struct PrepaintState {
-    scroll_offset: Point<Pixels>,
-    line: Option<ShapedLine>,
-    lines: Option<SmallVec<[WrappedLine; 1]>>,
-    cursor: Option<PaintQuad>,
-    selection: Option<PaintQuad>,
-    bounds: Bounds<Pixels>,
-}
-
-impl IntoElement for TextElement {
-    type Element = Self;
-
-    fn into_element(self) -> Self::Element {
-        self
-    }
-}
-
-impl Element for TextElement {
-    type RequestLayoutState = ();
-    type PrepaintState = PrepaintState;
-
-    fn id(&self) -> Option<ElementId> {
-        None
-    }
-
-    fn request_layout(
-        &mut self,
-        _id: Option<&GlobalElementId>,
-        cx: &mut WindowContext,
-    ) -> (LayoutId, Self::RequestLayoutState) {
-        let mut style = Style::default();
-        style.size.width = relative(1.).into();
-        style.size.height = if self.input.read(cx).multi_line {
-            (10. * cx.line_height()).into()
-        } else {
-            cx.line_height().into()
-        };
-        (cx.request_layout(style, []), ())
-    }
-
-    fn prepaint(
-        &mut self,
-        _id: Option<&GlobalElementId>,
-        bounds: Bounds<Pixels>,
-        _request_layout: &mut Self::RequestLayoutState,
-        cx: &mut WindowContext,
-    ) -> Self::PrepaintState {
-        let multi_line = self.input.read(cx).multi_line;
-        let input = self.input.read(cx);
-        let text = input.text.clone();
-        let placeholder = input.placeholder.clone();
-        let selected_range = input.selected_range.clone();
-        let cursor_offset = input.cursor_offset();
-        let style = cx.text_style();
-
-        let (display_text, text_color) = if text.is_empty() {
-            (placeholder, cx.theme().muted_foreground)
-        } else if input.masked {
-            (
-                "*".repeat(text.chars().count()).into(),
-                cx.theme().foreground,
-            )
-        } else {
-            (text, cx.theme().foreground)
-        };
-
-        let run = TextRun {
-            len: display_text.len(),
-            font: style.font(),
-            color: text_color,
-            background_color: None,
-            underline: None,
-            strikethrough: None,
-        };
-
-        let runs = if let Some(marked_range) = input.marked_range.as_ref() {
-            vec![
-                TextRun {
-                    len: marked_range.start,
-                    ..run.clone()
-                },
-                TextRun {
-                    len: marked_range.end - marked_range.start,
-                    underline: Some(UnderlineStyle {
-                        color: Some(run.color),
-                        thickness: px(1.0),
-                        wavy: false,
-                    }),
-                    ..run.clone()
-                },
-                TextRun {
-                    len: display_text.len() - marked_range.end,
-                    ..run.clone()
-                },
-            ]
-            .into_iter()
-            .filter(|run| run.len > 0)
-            .collect()
-        } else {
-            vec![run]
-        };
-
-        let font_size = style.font_size.to_pixels(cx.rem_size());
-        let mut line = None;
-        let mut lines = None;
-        if multi_line {
-            let wrap_width = bounds.size.width;
-            lines = cx
-                .text_system()
-                .shape_text(display_text, font_size, &runs, Some(wrap_width))
-                .ok();
-        } else {
-            line = cx
-                .text_system()
-                .shape_line(display_text, font_size, &runs)
-                .ok()
-        };
-
-        // Calculate the scroll offset to keep the cursor in view
-        let mut scroll_offset = input.scroll_offset;
-        let mut bounds = bounds;
-        let mut selection = None;
-        let mut cursor = None;
-        const RIGHT_MARGIN: Pixels = px(5.);
-        const INSET: Pixels = px(0.5);
-
-        if let Some(line) = &line {
-            let cursor_pos = line.x_for_index(cursor_offset);
-            let cursor_start = line.x_for_index(selected_range.start);
-            let cursor_end = line.x_for_index(selected_range.end);
-            let cursor_moved = input.last_cursor_offset != Some(cursor_offset);
-            let selection_changed = input.last_selected_range != Some(selected_range.clone());
-
-            if cursor_moved || selection_changed {
-                scroll_offset.x =
-                    if scroll_offset.x + cursor_pos > (bounds.size.width - RIGHT_MARGIN) {
-                        // cursor is out of right
-                        bounds.size.width - RIGHT_MARGIN - cursor_pos
-                    } else if scroll_offset.x + cursor_pos < px(0.) {
-                        // cursor is out of left
-                        scroll_offset.x - cursor_pos
-                    } else {
-                        scroll_offset.x
-                    };
-
-                if input.selection_reversed {
-                    if scroll_offset.x + cursor_start < px(0.) {
-                        // selection start is out of left
-                        scroll_offset.x = -cursor_start;
-                    }
-                } else {
-                    if scroll_offset.x + cursor_end <= px(0.) {
-                        // selection end is out of left
-                        scroll_offset.x = -cursor_end;
-                    }
-                }
-            }
-
-            bounds.origin = bounds.origin + scroll_offset;
-
-            if selected_range.is_empty() && input.show_cursor(cx) {
-                // cursor blink
-                cursor = Some(fill(
-                    Bounds::new(
-                        point(bounds.left() + cursor_pos, bounds.top() + INSET),
-                        size(px(2.), bounds.bottom() - bounds.top() - INSET * 2),
-                    ),
-                    crate::blue_500(),
-                ))
-            } else {
-                // selection background
-                selection = Some(fill(
-                    Bounds::from_corners(
-                        point(
-                            bounds.left() + line.x_for_index(selected_range.start),
-                            bounds.top(),
-                        ),
-                        point(
-                            bounds.left() + line.x_for_index(selected_range.end),
-                            bounds.bottom(),
-                        ),
-                    ),
-                    cx.theme().selection,
-                ))
-            };
-        }
-
-        PrepaintState {
-            scroll_offset,
-            bounds,
-            line,
-            lines,
-            cursor,
-            selection,
-        }
-    }
-
-    fn paint(
-        &mut self,
-        _id: Option<&GlobalElementId>,
-        _: Bounds<Pixels>,
-        _request_layout: &mut Self::RequestLayoutState,
-        prepaint: &mut Self::PrepaintState,
-        cx: &mut WindowContext,
-    ) {
-        let focus_handle = self.input.read(cx).focus_handle.clone();
-        let focused = focus_handle.is_focused(cx);
-        let bounds = prepaint.bounds;
-        let selected_range = self.input.read(cx).selected_range.clone();
-        let cursor = self.input.read(cx).cursor_offset();
-
-        cx.handle_input(
-            &focus_handle,
-            ElementInputHandler::new(bounds, self.input.clone()),
-        );
-        if let Some(selection) = prepaint.selection.take() {
-            cx.paint_quad(selection)
-        }
-
-        let mut last_layout = None;
-
-        // Paint single line text
-        if let Some(line) = prepaint.line.take() {
-            line.paint(bounds.origin, cx.line_height(), cx).unwrap();
-            last_layout = Some(line);
-        }
-
-        // Paint multi line text
-        if let Some(lines) = prepaint.lines.take() {
-            let line_height = cx.line_height();
-            let origin = bounds.origin;
-
-            for (i, line) in lines.into_iter().enumerate() {
-                let p = point(origin.x, origin.y + i as f32 * line_height);
-                line.paint(p, line_height, cx).unwrap();
-            }
-        }
-
-        if focused {
-            if let Some(cursor) = prepaint.cursor.take() {
-                cx.paint_quad(cursor);
-            }
-        }
-        self.input.update(cx, |input, _cx| {
-            input.scroll_offset = prepaint.scroll_offset;
-            // FIXME: To support multi-line text, we need to store the last layout for each line.
-            input.last_layout = last_layout;
-            input.last_bounds = Some(bounds);
-            input.last_cursor_offset = Some(cursor);
-            input.last_selected_range = Some(selected_range);
-        });
-
-        self.paint_mouse_listeners(cx);
-    }
-}
-
 impl Render for TextInput {
     fn render(&mut self, cx: &mut ViewContext<Self>) -> impl IntoElement {
         const LINE_HEIGHT: Rems = Rems(1.25);
@@ -1215,9 +941,7 @@ impl Render for TextInput {
                     .flex_grow()
                     .overflow_x_hidden()
                     .cursor_text()
-                    .child(TextElement {
-                        input: cx.view().clone(),
-                    }),
+                    .child(TextElement::new(cx.view().clone())),
             )
             .when(self.loading, |this| {
                 this.child(Indicator::new().color(cx.theme().muted_foreground))
