+++ crates/ui/src/input/mod.rs
@@ -1,6 +1,7 @@
 mod blink_cursor;
 mod change;
 mod clear_button;
+mod element;
 mod input;
 mod otp_input;
